+++ src/platforms/render/opengl/opengl_texture.hh
@@ -11,7 +11,7 @@ namespace Recursion::opengl::render
     class OpenGLTexture : public core::render::Texture
     {
     public:
-        OpenGLTexture(const std::string &path, uint32_t tile_factor = 1, bool is_transparent = false, uint32_t wrap_s = GL_REPEAT,
+        OpenGLTexture(const std::string &path, float tile_factor = 1, bool is_transparent = false, uint32_t wrap_s = GL_REPEAT,
                       uint32_t wrap_t = GL_REPEAT,
                       uint32_t min_filter = GL_LINEAR_MIPMAP_LINEAR,
                       uint32_t mag_filter = GL_LINEAR);
@@ -33,7 +33,7 @@ namespace Recursion::opengl::render
         static uint32_t TEXTURE_UNIT_LIMIT;
         static uint32_t AVAILABLE_TEXTURE_UNIT;
         bool is_transparent;
-        int32_t tile_factor;
+        float tile_factor;
 
     protected:
         int32_t width;
